animate & fix data component stuff

+++ src/main/java/me/alphamode/beta/proxy/networking/packet/modern/packets/s2c/play/S2CAnimatePacket.java
@@ -7,11 +7,11 @@ import me.alphamode.beta.proxy.util.codec.ModernStreamCodecs;
 import me.alphamode.beta.proxy.util.codec.StreamCodec;
 
 public record S2CAnimatePacket(int entityId, short action) implements S2CPlayPacket {
-	public static final int SWING_MAIN_HAND = 0;
-	public static final int WAKE_UP = 2;
-	public static final int SWING_OFF_HAND = 3;
-	public static final int CRITICAL_HIT = 4;
-	public static final int MAGIC_CRITICAL_HIT = 5;
+	public static final short SWING_MAIN_HAND = 0;
+	public static final short WAKE_UP = 2;
+	public static final short SWING_OFF_HAND = 3;
+	public static final short CRITICAL_HIT = 4;
+	public static final short MAGIC_CRITICAL_HIT = 5;
 
 	public static final StreamCodec<ByteBuf, S2CAnimatePacket> CODEC = StreamCodec.composite(
 			ModernStreamCodecs.VAR_INT,
